Rename beziers helper to bezier to avoid conflict with bezier.py
