+++ Lib/fontquant/helpers/bezier.py
